Corretta la classe Database Session dove in insert era Integer mentre noi usiamo Long per gli ID, tolti setter degli ID dalle classi e aggiunti ID autogenerati autoincrementali.

+++ Source/childcare-server/src/main/java/com/polimi/childcare/server/database/DatabaseSession.java
@@ -68,16 +68,16 @@ public class DatabaseSession
 
     //region Metodi per Entita
 
-    public <T> Integer insert(T element)
+    public <T> Long insert(T element)
     {
         if(sessionFactory != null) {
             Transaction tx = null;
-            Integer ID = null;
+            Long ID = null;
 
             try (Session session = sessionFactory.openSession())
             {
                 tx = session.beginTransaction();
-                ID = (Integer) session.save(element);
+                ID = (Long) session.save(element);
                 tx.commit();
             } catch (HibernateException e) {
                 if (tx != null) tx.rollback();
